Login and registration page with role designation
Created a series of HTML/PHP files that utilize a database named "users_db"  with a table named "users" with the following fields: id, name, email, password, and role.
To log in, you must use the email and password of a user, and you will then be redirected to user_page.php or admin_page.php, depending on the privilege level (role) of the account.  Logout functionality has also been included for future use.